+++ mirai-core/src/commonMain/kotlin/utils/retryWithServers.kt
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2020 Mamoe Technologies and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
+ *
+ * https://github.com/mamoe/mirai/blob/master/LICENSE
+ */
+
+package net.mamoe.mirai.internal.utils
+
+import kotlinx.coroutines.withTimeoutOrNull
+import kotlin.math.roundToInt
+
+
+internal suspend inline fun List<Pair<Int, Int>>.retryWithServers(
+    timeoutMillis: Long,
+    onFail: (exception: Throwable?) -> Unit,
+    crossinline block: suspend (ip: String, port: Int) -> Unit
+) {
+    require(this.isNotEmpty()) { "receiver of retryWithServers must not be empty" }
+
+    var exception: Throwable? = null
+    for (pair in this) {
+        return kotlin.runCatching {
+            withTimeoutOrNull(timeoutMillis) {
+                block(pair.first.toIpV4AddressString(), pair.second)
+            }
+        }.recover {
+            if (exception != null) {
+                exception!!.addSuppressed(it)
+            }
+            exception = it
+            null
+        }.getOrNull() ?: continue
+    }
+
+    onFail(exception)
+}
+
+internal fun Int.sizeToString() = this.toLong().sizeToString()
+internal fun Long.sizeToString(): String {
+    return if (this < 1024) {
+        "$this B"
+    } else ((this * 100.0 / 1024).roundToInt() / 100.0).toString() + " KiB"
+}
